source_codes/agentForge/requirements.txt
gradio
smolagents
python-dotenv
requests
duckduckgo_search
pandas
markdownify
pyyaml
smolagents[litellm]